requirements.txt
pytest==7.4.4
allure-pytest==2.13.2
requests==2.31.0
jsonschema